% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processLines.R
\name{processLines}
\alias{processLines}
\title{Process Line data}
\usage{
processLines(lineData, overWrite)
}
\arguments{
\item{lineData}{: Data frame containing all the measured line features.}

\item{overWrite}{: should this rewrite any existing line files. TRUE/FALSE}
}
\value{

}
\description{
Process Line data
}
