% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPolygons.R
\name{processPolygons}
\alias{processPolygons}
\title{Process polygon data}
\usage{
processPolygons(polyData, overWrite)
}
\arguments{
\item{polyData}{Data frame containing all the measured polygon features.}

\item{overWrite}{should this rewrite any existing line files. TRUE/FALSE}
}
\value{

}
\description{
takes an dataframe of points and produces polygons
}
