% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceMapsGoogle.R
\name{Produce maps}
\alias{Produce maps}
\alias{produceMapsGoogle}
\title{Generate Maps}
\usage{
produceMapsGoogle(line, polygons, zoomLevel, fileLocation, apiKey)
}
\arguments{
\item{line}{: sf object containing specific lines}

\item{polygons}{: sf object containing specific polygons}

\item{zoomLevel}{: zoom level for the map}

\item{fileLocation}{: where output will be saved too}

\item{apiKey}{: key for google account}
}
\value{

}
\description{
generates a map using google imagery
}
